feat: complete baseapi

+++ web-voyager-ddd-domain/src/main/java/io/github/sztuwebclub/webvoyager/domain/problem/IProblemRepo.java
@@ -7,7 +7,7 @@ public interface IProblemRepo {
 
     List<Problem> pageQuery(Integer page, Integer pagesize, String title);
 
-    Problem getProblemById(Integer id);
+    Problem getProblemById(Long id);
 
-    ProblemDetails getProblemDeatilsById(Integer id);
+    ProblemDetails getProblemDeatilsById(Long id);
 }
