+++ web-voyager-ddd-infra/src/main/java/io/github/sztuwebclub/webvoyager/infra/dao/ProblemDatailsMapper.java
@@ -8,5 +8,5 @@ import org.apache.ibatis.annotations.Select;
 
 public interface ProblemDatailsMapper extends BaseMapper<ProblemDetailsPO> {
     @Select("select id, description, hint_input, hint_output, sample_input, sample_output, hint, input, output from problem_details where id = #{id}")
-    ProblemDetails getProblemDeatilsById(@Param("id") Integer id);
+    ProblemDetails getProblemDeatilsById(@Param("id") Long id);
 }
